Move universe.v3.model decoders into companion objects
