+++ build.sbt
@@ -25,6 +25,7 @@ lazy val model = project.in(file("cosmos-model"))
     libraryDependencies ++=
       Deps.scalaUri
         ++ Deps.circeCore
+        ++ Deps.circe
         ++ Deps.twitterUtilCore
         ++ Deps.fastparse
         ++ Deps.scalaCheckShapeless
