+++ cosmos-json/src/main/scala/com/mesosphere/cosmos/rpc/v1/circe/Decoders.scala
@@ -7,7 +7,6 @@ import com.mesosphere.cosmos.thirdparty.marathon.circe.Decoders._
 import com.mesosphere.universe
 import com.mesosphere.universe.common.circe.Decoders._
 import com.mesosphere.universe.v2.circe.Decoders._
-import com.mesosphere.universe.v3.circe.Decoders._
 import io.circe._
 import io.circe.generic.semiauto._
 import scala.util.Left
