+++ cosmos-json/src/main/scala/com/mesosphere/cosmos/rpc/v2/circe/Decoders.scala
@@ -3,7 +3,6 @@ package com.mesosphere.cosmos.rpc.v2.circe
 import com.mesosphere.cosmos.rpc.v2.model.DescribeResponse
 import com.mesosphere.cosmos.rpc.v2.model.InstallResponse
 import com.mesosphere.cosmos.thirdparty.marathon.circe.Decoders._
-import com.mesosphere.universe.v3.circe.Decoders._
 import io.circe.Decoder
 import io.circe.generic.semiauto._
