+++ cosmos-json/src/main/scala/com/mesosphere/cosmos/rpc/v2/circe/Encoders.scala
@@ -1,7 +1,6 @@
 package com.mesosphere.cosmos.rpc.v2.circe
 
 import com.mesosphere.cosmos.thirdparty.marathon.circe.Encoders.encodeAppId
-import com.mesosphere.universe.v3.circe.Encoders._
 import com.mesosphere.cosmos.rpc
 import com.mesosphere.cosmos.rpc.v2.model.DescribeResponse
 import io.circe.Encoder
