+++ cosmos-json/src/test/scala/com/mesosphere/universe/v3/circe/DcosReleaseVersionEncoderDecoderSpec.scala
@@ -1,7 +1,5 @@
 package com.mesosphere.universe.v3.circe
 
-import com.mesosphere.universe.v3.circe.Decoders._
-import com.mesosphere.universe.v3.circe.Encoders._
 import com.mesosphere.universe.v3.model.DcosReleaseVersion
 import com.mesosphere.universe.v3.model.DcosReleaseVersion._
 import io.circe.Json
