+++ cosmos-model/src/main/scala/com/mesosphere/universe/common/ByteBuffers.scala
