+++ cosmos-model/src/main/scala/com/mesosphere/universe/common/circe/Encoders.scala
