+++ cosmos-model/src/main/scala/com/mesosphere/universe/v3/model/Container.scala
@@ -1,5 +1,15 @@
 package com.mesosphere.universe.v3.model
 
+import io.circe.Decoder
+import io.circe.Encoder
+import io.circe.generic.semiauto.deriveDecoder
+import io.circe.generic.semiauto.deriveEncoder
+
 case class Container(
   docker: Map[String, String]
 )
+
+object Container {
+  implicit val decodeContainer: Decoder[Container] = deriveDecoder[Container]
+  implicit val encodeContainer: Encoder[Container] = deriveEncoder[Container]
+}
