+++ cosmos-model/src/main/scala/com/mesosphere/universe/v3/model/ReleaseVersion.scala
@@ -1,8 +1,16 @@
 package com.mesosphere.universe.v3.model
 
+import cats.syntax.either._
 import com.twitter.util.Return
 import com.twitter.util.Throw
 import com.twitter.util.Try
+import io.circe.Decoder
+import io.circe.Encoder
+import io.circe.HCursor
+import io.circe.DecodingFailure
+import io.circe.syntax.EncoderOps
+import io.circe.generic.semiauto.deriveDecoder
+import io.circe.generic.semiauto.deriveEncoder
 
 final class ReleaseVersion private(val value: Long) extends AnyVal
 
@@ -20,5 +28,17 @@ object ReleaseVersion {
   implicit val packageDefinitionReleaseVersionOrdering: Ordering[ReleaseVersion] = {
     Ordering.by(_.value)
   }
+
+  implicit val encodePackageDefinitionReleaseVersion: Encoder[ReleaseVersion] = {
+    Encoder.instance(_.value.asJson)
+  }
+
+  implicit val decodePackageDefinitionReleaseVersion: Decoder[ReleaseVersion] =
+    Decoder.instance[ReleaseVersion] { (c: HCursor) =>
+      c.as[Long].map(ReleaseVersion(_)).flatMap {
+        case Return(v) => Right(v)
+        case Throw(e) => Left(DecodingFailure(e.getMessage, c.history))
+      }
+    }
 
 }
