+++ cosmos-model/src/main/scala/com/mesosphere/universe/v3/model/Resource.scala
@@ -1,5 +1,10 @@
 package com.mesosphere.universe.v3.model
 
+import io.circe.Decoder
+import io.circe.Encoder
+import io.circe.generic.semiauto.deriveDecoder
+import io.circe.generic.semiauto.deriveEncoder
+
 /**
   * Conforms to: https://universe.mesosphere.com/v3/schema/repo#/definitions/v20resource
   */
@@ -7,6 +12,11 @@ case class V2Resource(
   assets: Option[Assets] = None,
   images: Option[Images] = None
 )
+
+object V2Resource {
+  implicit val encodeV2Resource: Encoder[V2Resource] = deriveEncoder[V2Resource]
+  implicit val decodeV3V2Resource: Decoder[V2Resource] = deriveDecoder[V2Resource]
+}
 
 /**
   * Conforms to: https://universe.mesosphere.com/v3/schema/repo#/definitions/v30resource
@@ -16,3 +26,8 @@ case class V3Resource(
   images: Option[Images] = None,
   cli: Option[Cli] = None
 )
+
+object V3Resource {
+  implicit val encodeV3Resource: Encoder[V3Resource] = deriveEncoder[V3Resource]
+  implicit val decodeV3V3Resource: Decoder[V3Resource] = deriveDecoder[V3Resource]
+}
