+++ cosmos-model/src/main/scala/com/mesosphere/universe/v3/model/Tag.scala
@@ -1,9 +1,17 @@
 package com.mesosphere.universe.v3.model
 
+import cats.syntax.either._
 import com.twitter.util.Return
 import com.twitter.util.Throw
 import com.twitter.util.Try
 import java.util.regex.Pattern
+import io.circe.Decoder
+import io.circe.Encoder
+import io.circe.HCursor
+import io.circe.DecodingFailure
+import io.circe.generic.semiauto.deriveDecoder
+import io.circe.generic.semiauto.deriveEncoder
+import io.circe.syntax.EncoderOps
 
 final class Tag private(val value: String) extends AnyVal {
 
@@ -25,5 +33,19 @@ object Tag {
       ))
     }
   }
+
+  implicit val encodePackageDefinitionTag: Encoder[Tag] = {
+    Encoder.instance(_.value.asJson)
+  }
+
+  implicit val decodePackageDefinitionTag: Decoder[Tag] =
+    Decoder.instance[Tag] { (c: HCursor) =>
+      c.as[String].map(Tag(_)).flatMap {
+        case Return(r) => Right(r)
+        case Throw(ex) =>
+          val msg = ex.getMessage.replaceAllLiterally("assertion failed: ", "")
+          Left(DecodingFailure(msg, c.history))
+      }
+    }
 
 }
